+++ onedal/common/_policy.py
@@ -20,7 +20,7 @@ from onedal import _backend, _is_dpc_backend
 
 oneapi_is_available = "daal4py.oneapi" in sys.modules
 if oneapi_is_available:
-    from daal4py.oneapi import _get_sycl_ctxt, sycl_execution_context
+    from daal4py.oneapi import _get_sycl_ctxt
 
 
 def _get_policy(queue, *data):
@@ -69,10 +69,9 @@ if _is_dpc_backend:
         def __init__(self, queue):
             self._queue = queue
             self._d4p_interop = _Daal4PyContextReset()
-            if "sklearnex" in sys.modules:
-                from sklearnex._device_offload import DummySyclQueue
+            from daal4py.sklearn._device_offload import DummySyclQueue
 
-                if isinstance(queue, DummySyclQueue):
-                    super().__init__(self._queue.sycl_device.get_filter_string())
-                    return
+            if isinstance(queue, DummySyclQueue):
+                super().__init__(self._queue.sycl_device.get_filter_string())
+                return
             super().__init__(self._queue)
